First phase of mini Java Compiler
At this stage of syntactic analysis the Compiler is the driver class
which utilizes a Scanner class to transform character stream from a
source file into tokens. The tokens are then parsed according to the
mini Java grammar and source file is either accepted or rejected based
on exit code (0 accept; 4 reject).